config: the model-path and attention-backend decisions move out of the record
Six handlers, two families. `arg_groups/model_path_hook.py` resolves where the
weights come from (the HF/GGUF reference, the ModelScope snapshot download, the
object-storage URI). `arg_groups/attention_hook.py` holds the backend
compatibility rules, the linear-attention backend selection, and the
multi-item-scoring adjustment.

62 launch shapes report the same outcome as before, and the 24-shape
resolution dump is field-for-field identical.
