+++ python/sglang/srt/server_args.py
@@ -119,7 +119,7 @@ from sglang.srt.utils.common import (
 )
 from sglang.srt.utils.hf_transformers_utils import check_gguf_file
 from sglang.srt.utils.network import NetworkAddress, get_free_port, wait_port_available
-from sglang.srt.utils.runai_utils import ObjectStorageModel, is_runai_obj_uri
+from sglang.srt.utils.runai_utils import is_runai_obj_uri
 
 logger = logging.getLogger(__name__)
 
@@ -4236,23 +4236,9 @@ class ServerArgs:
             )
 
     def _handle_model_source_paths(self):
-        """Prepare metadata for model paths backed by remote object stores."""
-        cfg = resolving_view(self)
-        self._resolve_hf_gguf_model_path()
+        from sglang.srt.arg_groups.model_path_hook import handle_model_source_paths
 
-        seen_paths = set()
-        for model_path in (
-            cfg.model_path,
-            cfg.tokenizer_path,
-            cfg.speculative_draft_model_path,
-        ):
-            if (
-                model_path is not None
-                and model_path not in seen_paths
-                and is_runai_obj_uri(model_path)
-            ):
-                ObjectStorageModel.download_and_get_path(model_path)
-                seen_paths.add(model_path)
+        handle_model_source_paths(self)
 
     def _handle_pd_disaggregation(self):
         from sglang.srt.arg_groups.pd_disaggregation_hook import (
@@ -4518,78 +4504,9 @@ class ServerArgs:
             )
 
     def _handle_modelscope_paths(self):
-        """Resolve model / tokenizer / speculative-draft paths from the local
-        ModelScope cache when possible, falling back to snapshot_download
-        for any path that is not already present on disk.
+        from sglang.srt.arg_groups.model_path_hook import handle_modelscope_paths
 
-        Note: speculative_token_map is intentionally NOT handled here
-        because its value uses repo_id/filename semantics rather than a
-        plain repo ID.  That resolution lives in
-        :func:`sglang.srt.speculative.spec_utils.load_token_map`.
-        """
-        cfg = resolving_view(self)
-
-        ms_root = None
-        ms_snapshot_download = None
-
-        def _resolve_or_download(
-            path: Optional[str],
-            ignore_patterns: Optional[list] = None,
-            revision: Optional[str] = None,
-        ) -> Optional[str]:
-            nonlocal ms_root, ms_snapshot_download
-            if path is None:
-                return None
-            if not path or os.path.exists(path):
-                return path
-
-            if ms_snapshot_download is None:
-                from modelscope.hub.snapshot_download import (
-                    snapshot_download as _ms_snapshot_download,
-                )
-                from modelscope.utils.file_utils import get_model_cache_root
-
-                ms_snapshot_download = _ms_snapshot_download
-                ms_root = get_model_cache_root()
-
-            # Check ModelScope default cache
-            cached = os.path.join(ms_root, path)
-            if os.path.exists(cached):
-                return cached
-            # Check user-specified download dir
-            if cfg.download_dir:
-                alt = os.path.join(cfg.download_dir, path)
-                if os.path.exists(alt):
-                    return alt
-
-            # Cache miss — download from ModelScope hub
-            return ms_snapshot_download(
-                path,
-                cache_dir=cfg.download_dir,
-                revision=revision,
-                **({"ignore_patterns": ignore_patterns} if ignore_patterns else {}),
-            )
-
-        self._declare(
-            "_handle_modelscope_paths",
-            model_path=_resolve_or_download(cfg.model_path, revision=cfg.revision),
-        )
-        self._declare(
-            "_handle_modelscope_paths",
-            tokenizer_path=_resolve_or_download(
-                cfg.tokenizer_path,
-                ignore_patterns=["*.bin", "*.safetensors"],
-                revision=cfg.revision,
-            ),
-        )
-        if cfg.speculative_draft_model_path:
-            self._declare(
-                "_handle_modelscope_paths",
-                speculative_draft_model_path=_resolve_or_download(
-                    cfg.speculative_draft_model_path,
-                    revision=cfg.speculative_draft_model_revision or "main",
-                ),
-            )
+        handle_modelscope_paths(self)
 
     def _handle_hpu_backends(self):
         cfg = resolving_view(self)
@@ -4761,52 +4678,9 @@ class ServerArgs:
                 )
 
     def _handle_multi_item_scoring(self):
-        """Setup and validate multi-item scoring constraints.
+        from sglang.srt.arg_groups.attention_hook import handle_multi_item_scoring
 
-        Auto-disables settings incompatible with MIS mechanics (CUDA graph,
-        radix cache, chunked prefill). Asserts on attention backend since
-        changing it silently could surprise users who intentionally picked
-        a non-flashinfer backend.
-        """
-        cfg = resolving_view(self)
-        if not cfg.enable_mis:
-            return
-
-        if cfg.cuda_graph_config.decode.backend != Backend.DISABLED:
-            logger.warning("CUDA graph is disabled because --enable-mis is set.")
-        self._declare(
-            "_handle_multi_item_scoring",
-            cuda_graph_config=with_phase(
-                cfg.cuda_graph_config, Phase.DECODE, backend=Backend.DISABLED
-            ),
-        )
-        self._declare(
-            "_handle_multi_item_scoring",
-            cuda_graph_config=with_phase(
-                cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.DISABLED
-            ),
-        )
-
-        if not cfg.disable_radix_cache:
-            logger.warning("Radix cache is disabled because --enable-mis is set.")
-            self._declare(
-                "_handle_multi_item_scoring",
-                disable_radix_cache=True,
-            )
-
-        if cfg.chunked_prefill_size != -1:
-            logger.warning("Chunked prefill is disabled because --enable-mis is set.")
-            self._declare(
-                "_handle_multi_item_scoring",
-                chunked_prefill_size=-1,
-            )
-
-        prefill_backend, decode_backend = self._resolved_attention_backends()
-        assert prefill_backend == "flashinfer" and decode_backend == "flashinfer", (
-            "Multi-item scoring requires flashinfer attention backend for custom attention mask support. "
-            f"Please set --attention-backend flashinfer when using --enable-mis. "
-            f"Current backends: prefill={prefill_backend}, decode={decode_backend}"
-        )
+        handle_multi_item_scoring(self)
 
     def _handle_gpu_memory_settings(self, gpu_mem):
         """
@@ -5994,178 +5868,11 @@ class ServerArgs:
                 return "triton"
 
     def _handle_attention_backend_compatibility(self):
-        cfg = resolving_view(self)
-        model_config = self.get_model_config()
-
-        # The attention_backend write clusters of this handler moved to the
-        # resolution pipeline (arg_groups/overrides.py), each invoked below at
-        # its legacy slot; the interleaved non-attention adjustments stay.
-        from sglang.srt.arg_groups.overrides import (
-            _attention_backend_default,
-            _attention_backend_dual_chunk,
-            _attention_backend_fa3_fp8_fallback,
-            _attention_backend_platform_fallbacks,
-            _fa4_page_constraint,
-            _intel_xpu_page_constraint,
-            _mla_backend_page_constraints,
-            run_post_process_pass,
+        from sglang.srt.arg_groups.attention_hook import (
+            handle_attention_backend_compatibility,
         )
 
-        # Split-backend override + default fill.
-        run_post_process_pass(self, _attention_backend_default)
-
-        # Torch native and flex attention backends
-        attention_backend = resolved_view(self).attention_backend
-        if attention_backend == "torch_native":
-            logger.warning(
-                "Cuda graph is disabled because of using torch native attention backend"
-            )
-            self._declare(
-                "_handle_attention_backend_compatibility",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.DECODE, backend=Backend.DISABLED
-                ),
-            )
-            self._declare(
-                "_handle_attention_backend_compatibility",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.DISABLED
-                ),
-            )
-
-        if attention_backend == "flex_attention":
-            logger.warning(
-                "Cuda graph is disabled because of using torch Flex Attention backend"
-            )
-            self._declare(
-                "_handle_attention_backend_compatibility",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.DECODE, backend=Backend.DISABLED
-                ),
-            )
-            self._declare(
-                "_handle_attention_backend_compatibility",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.DISABLED
-                ),
-            )
-            assert (
-                cfg.speculative_algorithm is None
-            ), "Speculative decoding is currently not supported with Flex Attention backend"
-
-        # Whisper's encoder token padding conflicts with prefix caching.
-        # Only disable for Whisper; other encoder-decoder models (e.g., mllama) use radix cache.
-        if (
-            model_config.is_encoder_decoder
-            and not cfg.disable_radix_cache
-            and "WhisperForConditionalGeneration"
-            in (model_config.hf_config.architectures or [])
-        ):
-            logger.info("Radix cache is disabled for Whisper")
-            self._declare(
-                "_handle_attention_backend_compatibility",
-                disable_radix_cache=True,
-            )
-
-        # Major NVIDIA platforms backends: the page-size snaps of this family
-        # moved to the resolution pipeline (arg_groups/overrides.py:
-        # _mla_backend_page_constraints); the raises and the cutedsl prefill
-        # fallback stay below.
-        run_post_process_pass(self, _mla_backend_page_constraints)
-
-        # The TRT-LLM / tokenspeed MLA kv-dtype validations moved to the
-        # resolution pipeline (arg_groups/overrides.py:
-        # _mla_kv_cache_dtype_checks), invoked here at their legacy slot.
-        from sglang.srt.arg_groups.overrides import _mla_kv_cache_dtype_checks
-
-        run_post_process_pass(self, _mla_kv_cache_dtype_checks)
-
-        # The CuteDSL MLA validation + prefill fill moved to the resolution
-        # pipeline (arg_groups/overrides.py: _cutedsl_prefill_backend_fill),
-        # invoked here at its legacy slot.
-        from sglang.srt.arg_groups.overrides import _cutedsl_prefill_backend_fill
-
-        run_post_process_pass(self, _cutedsl_prefill_backend_fill)
-
-        prefill_backend, decode_backend = self._resolved_attention_backends()
-        if "trtllm_mha" in (prefill_backend, decode_backend):
-            if prefill_backend == "trtllm_mha" and not (
-                is_sm90_supported() or is_sm100_supported() or is_sm120_supported()
-            ):
-                raise ValueError(
-                    "TRTLLM MHA backend for prefill requires Hopper (SM90), Blackwell (SM100), or SM120 GPUs. "
-                    "Please use a different prefill backend."
-                )
-            if (
-                prefill_backend == "trtllm_mha"
-                and is_sm120_supported()
-                and (
-                    cfg.kv_cache_dtype == "fp8_e4m3"
-                    or (
-                        envs.SGLANG_SKIP_SOFTMAX_PREFILL_THRESHOLD_SCALE_FACTOR.get()
-                        or 0.0
-                    )
-                    > 0
-                )
-            ):
-                raise ValueError(
-                    "TRTLLM FMHAv2 prefill on SM120 does not support "
-                    "fp8_e4m3 KV cache or skip-softmax."
-                )
-            if decode_backend == "trtllm_mha" and not (
-                is_sm90_supported() or is_sm100_supported() or is_sm120_supported()
-            ):
-                raise ValueError(
-                    "TRTLLM MHA backend for decode is only supported on Hopper (SM90), Blackwell (SM100) and (SM120) GPUs. Please use a different decode backend."
-                )
-            if (
-                prefill_backend == "trtllm_mha"
-                and not is_sm100_supported()
-                and (cfg.enable_prefill_context_parallel or cfg.attn_cp_size > 1)
-            ):
-                raise ValueError(
-                    "Prefill context parallelism with the TRTLLM MHA prefill backend "
-                    "requires SM100 (trtllm-gen context kernel): the SM90/SM120 "
-                    "fmha_v2 prefill path does not implement CP shard masking."
-                )
-
-        run_post_process_pass(self, _attention_backend_fa3_fp8_fallback)
-
-        run_post_process_pass(self, _fa4_page_constraint)
-
-        # AMD platforms backends
-        if resolved_view(self).attention_backend == "aiter":
-            if model_config.context_len > 8192:
-                self._declare(
-                    "_handle_attention_backend_compatibility",
-                    mem_fraction_static=cfg.mem_fraction_static * 0.85,
-                )
-
-        # Other platforms backends
-        run_post_process_pass(self, _attention_backend_platform_fallbacks)
-
-        prefill_backend, decode_backend = self._resolved_attention_backends()
-        if self.use_mla_backend() and prefill_backend == "intel_xpu":
-            raise ValueError(
-                "intel_xpu backend is only supported on decode for MLA models, please set --decode-attention-backend to intel_xpu and do not set --attention-backend or --prefill-attention-backend to intel_xpu for prefill instead use triton."
-            )
-
-        run_post_process_pass(self, _intel_xpu_page_constraint)
-
-        # Dual chunk flash attention backend
-        run_post_process_pass(self, _attention_backend_dual_chunk)
-        if resolved_view(self).attention_backend == "dual_chunk_flash_attn":
-            logger.warning(
-                "Mixed chunk and radix cache are disabled when using dual-chunk flash attention backend"
-            )
-            self._declare(
-                "_handle_attention_backend_compatibility",
-                enable_mixed_chunk=False,
-            )
-            self._declare(
-                "_handle_attention_backend_compatibility",
-                disable_radix_cache=True,
-            )
+        handle_attention_backend_compatibility(self)
 
     def _handle_mxfp8_kv_cache_compatibility(self):
         from sglang.srt.arg_groups.kv_cache_hook import (
@@ -6215,233 +5922,9 @@ class ServerArgs:
         handle_int8_mamba_checkpoint(self)
 
     def _handle_linear_attn_backend(self):
-        cfg = resolving_view(self)
-        import torch
+        from sglang.srt.arg_groups.attention_hook import handle_linear_attn_backend
 
-        # SM100+: default to FlashInfer GDN decode (and MTP verify, via pool API)
-        # when the user hasn't explicitly chosen a decode backend and
-        # mamba-ssm-dtype is bf16 (required by FlashInfer GDN on SM100+).
-        # Fixed in FlashInfer v0.6.7: flashinfer-ai/flashinfer#2810
-        if (
-            cfg.linear_attn_decode_backend is None
-            and cfg.linear_attn_backend != "helion"
-            and is_sm100_supported()
-            and cfg.mamba_ssm_dtype == "bfloat16"
-            # Stage 4: flashinfer's recurrent_kda compiles the state slot stride
-            # as a free int64, so it reads the page-major/unified envelope-strided
-            # state correctly — the unified-memory skip is no longer needed (the
-            # page-major gate now allows flashinfer for linear-attn decode).
-        ):
-            self._declare(
-                "_handle_linear_attn_backend",
-                linear_attn_decode_backend="flashinfer",
-            )
-            logger.info(
-                "SM100+ detected with mamba-ssm-dtype=bfloat16, "
-                "defaulting --linear-attn-decode-backend to flashinfer."
-            )
-
-        # SM100+ FlashInfer GDN decode requires bf16 state; SM90 uses float32.
-        decode = cfg.linear_attn_decode_backend or cfg.linear_attn_backend
-
-        # FlashKDA is a prefill-only KDA kernel (no decode kernel) but shares the
-        # backend choice list, so guard it from being selected for decode: error
-        # on an explicit --linear-attn-decode-backend flashkda, and fall back to
-        # triton decode when it was only inherited from base=flashkda (prefill
-        # keeps FlashKDA).
-        if decode == "flashkda":
-            if cfg.linear_attn_decode_backend == "flashkda":
-                raise ValueError(
-                    "--linear-attn-decode-backend flashkda is not supported: "
-                    "FlashKDA is prefill-only. Use "
-                    "--linear-attn-prefill-backend flashkda (decode stays on triton)."
-                )
-            self._declare(
-                "_handle_linear_attn_backend",
-                linear_attn_decode_backend="triton",
-            )
-            decode = "triton"
-            logger.info(
-                "FlashKDA is prefill-only; using triton for KDA decode "
-                "(FlashKDA stays on prefill)."
-            )
-
-        if (
-            decode == "flashinfer"
-            and cfg.mamba_ssm_dtype != "bfloat16"
-            and is_cuda()
-            and torch.cuda.get_device_capability()[0] >= 10
-        ):
-            raise ValueError(
-                "--linear-attn-decode-backend flashinfer on SM100+ requires "
-                "--mamba-ssm-dtype bfloat16, "
-                f"got {cfg.mamba_ssm_dtype!r}"
-            )
-
-        verify = cfg.linear_attn_verify_backend
-        if verify is None and decode == "flashinfer":
-            verify = "flashinfer"
-        if (
-            verify == "flashinfer"
-            and cfg.mamba_ssm_dtype != "bfloat16"
-            and is_cuda()
-            and torch.cuda.get_device_capability()[0] >= 10
-        ):
-            raise ValueError(
-                "--linear-attn-verify-backend flashinfer on SM100+ requires "
-                "--mamba-ssm-dtype bfloat16, "
-                f"got {cfg.mamba_ssm_dtype!r}"
-            )
-
-        # SM100+ FlashInfer GDN prefill requires CUDA 13+ (CuTe DSL kernel)
-        # for correctness and best performance.
-        prefill = cfg.linear_attn_prefill_backend or cfg.linear_attn_backend
-        cuda_version = torch.version.cuda
-        cuda_major = int(cuda_version.split(".")[0]) if cuda_version is not None else 0
-        if (
-            prefill == "flashinfer"
-            and is_cuda()
-            and torch.cuda.get_device_capability()[0] >= 10
-            and cuda_major < 13
-        ):
-            raise ValueError(
-                "--linear-attn-prefill-backend flashinfer on SM100+ requires CUDA 13+, "
-                f"got CUDA {cuda_version or 'unknown'}"
-            )
-
-        # ReplaySSM buffered decode guards. Runs on Triton, or Helion for KDA.
-        # cuda-graph is supported (slice 1b: CUDA-graph-safe static
-        # write-cursor buffers). The RADIX prefix cache is now supported (slice
-        # 2b: the decode kernel force-flushes the ring into temporal[slot] on
-        # the radix track boundary `seq_lens % mamba_track_interval == 0`, and
-        # the COW copy-into-slot path resets the ring cursor) -- so the
-        # --disable-radix-cache requirement is dropped.
-        #
-        # Slice 2b only wires the no_buffer mamba scheduler strategy (the
-        # default). The extra_buffer strategy donates the track snapshot via
-        # `donate_mamba_ping_pong_slot` with a separate ping-pong slot swap that
-        # does NOT route through MambaPool.copy_from, so the ReplaySSM ring
-        # cursor of the donated/kept slot would not be reset there. Handling
-        # that donation path is a follow-up; for now require no_buffer.
-        if cfg.enable_linear_replayssm:
-            if decode not in {"triton", "helion"}:
-                raise ValueError(
-                    "--enable-linear-replayssm requires Triton, or Helion for "
-                    "KDA, as the linear-attn decode backend; got "
-                    f"--linear-attn-decode-backend={decode!r}."
-                )
-            from sglang.srt.arg_groups.overrides import (
-                mamba_extra_buffer_of,
-            )
-
-            if mamba_extra_buffer_of(resolved_view(self)):
-                raise ValueError(
-                    "--enable-linear-replayssm requires --mamba-radix-cache-strategy "
-                    "no_buffer (the default); the extra_buffer ping-pong "
-                    "donation path is not yet supported (follow-up). Got "
-                    f"--mamba-radix-cache-strategy={cfg.mamba_radix_cache_strategy!r}."
-                )
-            if cfg.disaggregation_mode != "null":
-                # The disaggregated decode pool (HybridMambaDecodeReqToTokenPool)
-                # is not wired for the ReplaySSM ring, so the flag would silently
-                # no-op there; disagg also runs a different cache/coordination
-                # flow that is not yet validated for ReplaySSM (follow-up).
-                raise ValueError(
-                    "--enable-linear-replayssm is not supported under PD "
-                    "disaggregation yet (follow-up). Got "
-                    f"--disaggregation-mode={cfg.disaggregation_mode!r}."
-                )
-            if cfg.linear_replayssm_cache_len < 1:
-                raise ValueError(
-                    "--linear-replayssm-cache-len must be >= 1, got "
-                    f"{cfg.linear_replayssm_cache_len}."
-                )
-
-        # ReplaySSM spec-verify (Part B of #28511): linear-chain target verify via
-        # fold-every-commit -- the verify stores each draft step's raw inputs into
-        # the per-slot (rawv, rawk, g, beta) window and the commit replays the
-        # accepted prefix into the fp32 checkpoint. The intra-window interaction
-        # uses a strictly-lower causal mask, so it is valid ONLY for a linear
-        # draft chain (speculative_eagle_topk in {None, 1}, i.e. NEXTN / MTP);
-        # EAGLE tree verify (topk > 1) must fall back to the recurrent verify.
-        # GDN sizes the window to the draft maximum; KDA (kda_backend) keeps a
-        # --linear-replayssm-cache-len window and folds via its own fused
-        # verify ring-write + commit_kda_replayssm_after_verify.
-        if cfg.enable_linear_replayssm_spec:
-            if cfg.speculative_eagle_topk not in (None, 1):
-                raise ValueError(
-                    "--enable-linear-replayssm-spec requires a linear draft chain "
-                    "(--speculative-eagle-topk in {None, 1}); the chunked verify "
-                    "kernel uses a strictly-lower causal mask and is invalid for "
-                    "EAGLE tree verify. Got "
-                    f"--speculative-eagle-topk={cfg.speculative_eagle_topk!r}."
-                )
-            if decode not in ("triton", "flashinfer"):
-                raise ValueError(
-                    "--enable-linear-replayssm-spec requires the triton or "
-                    "flashinfer linear-attn decode backend, got "
-                    f"--linear-attn-decode-backend={decode!r}."
-                )
-            from sglang.srt.speculative.ragged_verify import (
-                RaggedVerifyMode,
-                read_ragged_verify_mode,
-            )
-
-            ragged_mode = read_ragged_verify_mode()
-            if ragged_mode is not RaggedVerifyMode.STATIC:
-                # Ragged ring-writes need the KDA fold-every-commit family
-                # (DSPARK/DFLASH) + the triton verify kernel (nv_cutedsl falls
-                # back to it for ragged layouts). The GDN ring-write kernels do
-                # not take the ragged layout and the flashinfer verify kernel
-                # never writes the ring -> a stale ring would be folded; keep
-                # refusing those combinations.
-                _algo = (cfg.speculative_algorithm or "").upper()
-                verify = cfg.linear_attn_verify_backend
-                if _algo not in ("DSPARK", "DFLASH") or verify not in (
-                    "triton",
-                    "nv_cutedsl",
-                ):
-                    raise ValueError(
-                        "--enable-linear-replayssm-spec with "
-                        f"SGLANG_RAGGED_VERIFY_MODE={ragged_mode.value} requires the "
-                        "KDA fold-every-commit family (DSPARK/DFLASH) and a "
-                        "ring-writing verify kernel (--linear-attn-verify-backend "
-                        "triton or nv_cutedsl); got "
-                        f"algorithm={cfg.speculative_algorithm!r}, "
-                        f"verify={verify!r}. Use SGLANG_RAGGED_VERIFY_MODE=static."
-                    )
-            if cfg.disaggregation_mode == "prefill":
-                raise ValueError(
-                    "--enable-linear-replayssm-spec is not supported on a PD "
-                    "prefill server: the ring is spec-verify-only scratch and "
-                    "the prefill server never runs spec verify."
-                )
-            if cfg.enable_linear_replayssm:
-                raise ValueError(
-                    "--enable-linear-replayssm-spec and --enable-linear-replayssm are "
-                    "mutually exclusive: they share the ring storage but drive it "
-                    "with incompatible cursor protocols (per-decode-forward vs "
-                    "per-verify-commit advance)."
-                )
-            if cfg.mamba_ssm_dtype is None:
-                logger.info(
-                    "--enable-linear-replayssm-spec: setting --mamba-ssm-dtype "
-                    "float32 (the closed-loop exact fold keeps the SSM checkpoint "
-                    "bit-identical to the recurrent baseline)."
-                )
-                self._declare(
-                    "_handle_linear_attn_backend",
-                    mamba_ssm_dtype="float32",
-                )
-            elif cfg.mamba_ssm_dtype != "float32":
-                logger.warning(
-                    "--enable-linear-replayssm-spec with --mamba-ssm-dtype=%s: the "
-                    "closed-loop fold re-quantizes the committed state each "
-                    "commit/flush (fp32 keeps it bit-exact to the fp32 recurrent "
-                    "baseline), so it may drift over long sequences. Validate "
-                    "accuracy for your model.",
-                    cfg.mamba_ssm_dtype,
-                )
+        handle_linear_attn_backend(self)
 
     def _handle_legacy_cp_arguments(self):
         cfg = resolving_view(self)
@@ -7465,40 +6948,9 @@ class ServerArgs:
         resolve_storage_layout_compatibility(self)
 
     def _resolve_hf_gguf_model_path(self):
-        """Turn a Hub reference to a .gguf into a local file path."""
-        cfg = resolving_view(self)
-        from sglang.srt.utils.hf_transformers_utils import resolve_hf_gguf_reference
+        from sglang.srt.arg_groups.model_path_hook import resolve_hf_gguf_model_path
 
-        resolved = resolve_hf_gguf_reference(cfg.model_path, revision=cfg.revision)
-        if resolved is not None:
-            logger.info("Resolved GGUF %s -> %s", cfg.model_path, resolved)
-            if cfg.tokenizer_path == cfg.model_path:
-                self._declare(
-                    "_resolve_hf_gguf_model_path",
-                    tokenizer_path=resolved,
-                )
-            self._declare(
-                "_resolve_hf_gguf_model_path",
-                model_path=resolved,
-            )
-
-        # A speculative draft can be a .gguf too, and it is loaded by path, so it
-        # needs the same Hub-reference resolution as the target.
-        if cfg.speculative_draft_model_path:
-            resolved_draft = resolve_hf_gguf_reference(
-                cfg.speculative_draft_model_path,
-                revision=cfg.speculative_draft_model_revision,
-            )
-            if resolved_draft is not None:
-                logger.info(
-                    "Resolved draft GGUF %s -> %s",
-                    cfg.speculative_draft_model_path,
-                    resolved_draft,
-                )
-                self._declare(
-                    "_resolve_hf_gguf_model_path",
-                    speculative_draft_model_path=resolved_draft,
-                )
+        resolve_hf_gguf_model_path(self)
 
     def _handle_expert_pack(self):
         from sglang.srt.arg_groups.expert_pack_hook import handle_expert_pack
